refactor(kernel): phase 3a — extract accessor / lifecycle impl block to kernel::accessors
Move the first `impl LibreFangKernel { ... }` block (formerly mod.rs
lines 1106..2598, ~1500 lines, ~85 methods — accessors like `home_dir`,
`config_ref`, `cron`, `pairing_ref`, plus operational helpers like
`spawn_key_validation`, `spawn_approval_sweep_task`, `record`,
`auto_dream_*`, etc.) into `crates/librefang-kernel/src/kernel/accessors.rs`.

Sibling submodule of `kernel::mod`, so retains private-field/method
access into `LibreFangKernel` without any visibility surgery (same
pattern as the Phase 1 trait-impl moves).

Mechanical, behavior-preserving move — method bodies untouched.
mod.rs drops from ~18.2k to ~16.7k lines (-1500). 4 inherent
`impl LibreFangKernel` blocks remain in mod.rs (down from 5);
phase 3b/3c will continue chipping.

Verified: cargo check / clippy -- -D warnings / cargo check --tests
/ rustfmt --check all green.
